daily_audio_briefing/requirements.txt
gTTS
kokoro-onnx
soundfile
numpy
customtkinter
scrapetube
youtube-transcript-api
google-generativeai
python-dotenv
dateparser
qrcode
Pillow
